* 'master' of scm.dev.nokia.troll.no:qt/qt-symbian-staging: (35 commits)
  runonphone timestamps nanosecond formating set to 9 chars, 0 padded
  Put Qt translation files to rom on Symbian.
  Fix QDeclarativeEngine::setOfflineStoragePath() for Symbian
  Converting from double to qreal in gui
  Fix XML schema validation failure.
  Fix construction races in QtNetwork
  Removing accidental use of double instead of qreal from QLine
  Fix incorrect QFileInfo permissions on windows
  Converting accidental use of doubles to qreal in declarative
  Copy and Cut are not available for some QML editors
  Ignore CcpuCan calls if input context is being destroyed
  Remove one local variable from QCoeFepInputContext
  Symbian: Fix native dialog with splitscreen VKB focus loss issue
  Create auto test for http HEAD request
  QNAM - fix poor performance of HEAD request with authentication
  runonphone - Implement CODA autodetection
  Symbian: Parse ts filenames case-insensitively for localize_deployment
  Do not crash in copy() of pixmaps without an underlying bitmap
  QTBUG-9113 - reinstating tst_qobjectrace tests for Symbian on Qt 4.8
  symbian - document behaviour of QFile::handle()
  ...

+++ src/declarative/util/qdeclarativexmllistmodel.cpp
@@ -530,7 +530,7 @@ public:
 
     void notifyQueryStarted(bool remoteSource) {
         Q_Q(QDeclarativeXmlListModel);
-        progress = remoteSource ? 0.0 : 1.0;
+        progress = remoteSource ? qreal(0.0) : qreal(1.0);
         status = QDeclarativeXmlListModel::Loading;
         errorString.clear();
         emit q->progressChanged(progress);
